+++ CHANGELOG.md
@@ -122,6 +122,7 @@
 - Account MOVE Line:Add new field originDate.On FEC move line export,use the originDate instead of date field.
 - ManufOrder : Add new many-to-one field 'unit' and set its value from billOfMaterial unit and also added to report.
 - Tracking Number : change in stock move and stock move line based on tracking number configuration.
+- StockMove : Include sub stock location's lines in Report. 
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.